+++ CHANGELOG.md
@@ -47,7 +47,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Updated `gcpy/__init__.py` to account for files that have been added or moved
 - Moved `GC_72_vertical_levels.csv` from `benchmark/modules/` to `benchmark/config/`
 - Moved `benchmark/modules/species_database.yml` to `benchmark/config/species_database.yml`
-- Moved benchmark python scripts `gcpy/` to  the `benchmark/modules/` folder
+- Moved benchmark python scripts from `gcpy/` to  the `benchmark/modules/` folder
 
 ### Fixed
 - Generalized test for GCHP or GCClassic restart file in `regrid_restart_file.py`